+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - devcontainer for easier testing during development
 
+### Changed
+- Use `DataProviderToolkit` and data blocks to pre-validate and assemble entities
+
 
 ## [0.4.0] - 2025-08-28
 ### Fixed